Asked Claude for review comments.  It found some leaks, a few chunks of dead code, some stuff that was just sloppy, and a few uninitialized variable instances that worked out of luck until now.
Claude also proposed a time-domain polyphase resampler for common IF
to audio resample blocks.  They may get integrated later if the
performance is compelling. The real utility of a polyphase resampler
is its ability to handle short buffers. The new arrangements for Qt
audio in and out of the radio baseband may suggest smaller IF
buffers. (This will also improve some of the operational oddities
caused by long delay through the RF to audio path.)

+++ Signals/include/Utilities.hxx
@@ -49,7 +49,7 @@ namespace SoDa {
 
     auto vsize = (v0.size() > v1.size()) ? v1.size() : v0.size();    
 
-    for(int i = 0; i < vsize; i++) {
+    for(size_t i = 0; i < vsize; i++) {
       result += v0[i] * std::conj(v1[i]); 
     }
